Updated tests to wm beta 01, added a new one and improved both.
